*Changed a few more usage instances of single precision floating point to double precision. *Improved "Song>Create preview audio" to store the start and end times of the preview as INI tags (preview_start_time and preview_end_time) in the project.  Phase Shift currently supports using that start tag.  If chart audio is loaded, EOF will offer to create preview audio files using the given time range of the audio.  Subsequent calls to this function will read existing start and time stamps from the INI settings and use them to fill out the input fields.

+++ src/main.c
@@ -1045,7 +1045,7 @@ long eof_get_previous_note(long cnote)
 int eof_note_is_hopo(unsigned long cnote)
 {
 	double delta;
-	float hopo_delta = eof_song->tags->eighth_note_hopo ? 250.0 : 170.0;	//The proximity threshold to determine HOPO eligibility, when the tempo is 120BPM (within 1/3 beat, or 1/2 beat if 8th note HOPO option is enabled)
+	double hopo_delta = eof_song->tags->eighth_note_hopo ? 250.0 : 170.0;	//The proximity threshold to determine HOPO eligibility, when the tempo is 120BPM (within 1/3 beat, or 1/2 beat if 8th note HOPO option is enabled)
 	unsigned long i, cnotepos;
 	long pnote;
 	long beat = -1;
